+++ PhysicalFit/Models/TreadmillRecordDetails.cs
@@ -18,7 +18,7 @@ namespace PhysicalFit.Models
         public int DetectionTrainingRecordId { get; set; }
         public string IntenPercen { get; set; }
         public string MaxRunningSpeed { get; set; }
-        public int ForceDuration { get; set; }
+        public double ForceDuration { get; set; }
         public double Speed { get; set; }
         public Nullable<System.DateTime> CreatedDate { get; set; }
         public Nullable<System.DateTime> ModifiedDate { get; set; }
